create geojson with all properties + iri, use this instead of conflate script
